# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: World of Shadows — Backend API
  version: 1.0.0
  description: 'Machine-readable inventory of Flask `/api/v1` routes. Human-readable request/response examples: `docs/api/REFERENCE.md`.
    Tag taxonomy: `docs/api/openapi-taxonomy.md`.'
servers:
- url: /
  description: Same origin as backend (set base URL in your client).
...
paths:
...
  /api/v1/improvement/experiments/run:
    post:
      tags:
      - Improvement
      summary: POST /improvement/experiments/run
      operationId: post_api_v1_improvement_experiments_run
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/improvement/recommendations:
    get:
      tags:
      - Improvement
      summary: GET /improvement/recommendations
      operationId: get_api_v1_improvement_recommendations
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/improvement/recommendations/{package_id}/decision:
    post:
      tags:
      - Improvement
      summary: POST /improvement/recommendations/{package_id}/decision
      operationId: post_api_v1_improvement_recommendations_package_id_decision
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - name: package_id
        in: path
        required: true
        schema:
          type: string
  /api/v1/improvement/variants:
    post:
      tags:
      - Improvement
      summary: POST /improvement/variants
      operationId: post_api_v1_improvement_variants
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Flask-JWT-Extended access token from `POST /api/v1/auth/login`.
security:
- BearerAuth: []